+++ web/src/App.tsx
@@ -9,6 +9,7 @@ import { AacLibraryPage } from './AacLibraryPage.tsx';
 import { WorkerTokensPage } from './WorkerTokensPage.tsx';
 import { DashboardPage } from './DashboardPage.tsx';
 import { AuditLogPage } from './AuditLogPage.tsx';
+import { AccountPage } from './AccountPage.tsx';
 import { ConceptProposalsPage } from './ConceptProposalsPage.tsx';
 import { VerifyEmailPage } from './VerifyEmailPage.tsx';
 import { VerificationBanner } from './VerificationBanner.tsx';
@@ -205,6 +206,20 @@ export function App({
       <>
         {banner}
         <WorkerTokensPage
+          api={api}
+          account={account}
+          onLogout={() => void handleLogout()}
+          onNavigate={setView}
+        />
+      </>
+    );
+  }
+
+  if (view === 'account') {
+    return (
+      <>
+        {banner}
+        <AccountPage
           api={api}
           account={account}
           onLogout={() => void handleLogout()}